2_Neural_Network_Basics/Project_digit_recognizer_app/requirements.txt
Flask
tensorflow>=2.16
Pillow
numpy
gunicorn
huggingface_hub
keras